ridiculousness minimization
Long ago, autoconf was completely ridiculous in the
somersaults required to correctly configure check
stuff and it made some sense to generate all the
boilerplate parts of config macros when you had lots
of config macros.  Autoconf has now improved to the
point where it is merely ridiculous, making the
generation of autoconfigury boilerplate completely
ridiculous.  So, in the spirit of ridiculousness
reduction, this patch removes generated autoconfigury.

* config/confmacs.tlib: deleted
* config/conftest.tpl: deleted
* config/libopts.def: deleted
* config/m4.tgz: deleted
* config/misc.def: deleted
* config/ag_macros.m4: formerly generated
* config/libopts.m4: formerly generated
* Makefile.am: do not distribute deleted files
* config/.gitignore: do not ignore formerly generated files
* config/bootstrap.local: do not generate autoconf stuff