build reproducibility
If there are concurrent autogens running producing files of the same
name, it can choke on rename(2). It's a user error, so fail with
a comprehensible error message.

SOURCE_DATE_EPOCH is the newly current preferred method for setting
the date and time of the build products. If it is in use, then
in the man page templates, both use it and presume that the GNU
version of date is in use. This will fail if it is set and the
platform's date command is not GNU. Handle that failure.

* agen5/agDep.c (tidy_dep_file): handle rename(3) failure
* agen5/expOutput.c (ag_scm_out_move): handle rename(3) failure
* autoopts/tpl/agman-cmd.tpl (man page date): use SOURCE_DATE_EPOCH
* autoopts/tpl/agman3.tpl (man page date): use SOURCE_DATE_EPOCH
* autoopts/tpl/optexits.def: new file listing standard exit codes
  formatted for inclusion with option definition files
* config/bootstrap: disable temp directories under xtrace